Implemented and enabled user & kernel preemption.
* Kernel preemption:
At the end of EL1 interrupt handling, the kernel should check if
the current thread should be switched out (e.g. its time slice is
used up). If yes, call the schedule() to switch to the next thread.

* User preemption:
When a user process takes exceptions (system calls, interrupts, etc.),
the control is handed over to the exception handler. Before returning
to the original execution, the kernel can call the scheduler to do
a context switch to achieve user preemption.

Signed-off-by: Marco Wang <m.aesophor@gmail.com>