mm/mmu,VirtualMemoryMap: introduce PAGE_* macros and PD_PAGE_MASK
This commit introduces PAGE_SHIFT, PAGE_SIZE and PAGE_MASK macros,
as well as a special page mask called "PD_PAGE_MASK" which is meant for
extracting the physical address from a page descriptor.

The reason why we need a special page mask fro a page descriptor is:
when we are manipulating a page descriptor, the bits [58:55] are reserved for
software use, so we shouldn't use the regular PAGE_MASK. Instead, we define
a special "physical" page mask which extracts the physical address from
the bits [47:12] of a page descriptor.

Signed-off-by: Marco Wang <m.aesophor@gmail.com>